+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Gdf map visualization accepts both dataset and gdf. [#55](https://github.com/IN-CORE/pyincore-viz/issues/55)
 ## [1.6.0] - 2022-06-29
 ### Added
 - Household income and poverty visualization to pop result table. [#55](https://github.com/IN-CORE/pyincore-viz/issues/55)