+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
+- Gdf map visualization accepts both dataset and gdf. [#72](https://github.com/IN-CORE/pyincore-viz/issues/72)
 - Updated network visualization. [#68](https://github.com/IN-CORE/pyincore-viz/issues/68)
 
 ## [1.6.0] - 2022-06-29